PoC: Add SSL settings patching provided setup_irods.py
This seems to work. `icat` and `ires` start. Running `izonereport` seems
to work as well.

This patch is not complete. Adding support for the provided json should
be done as well:

https://github.com/irods/irods/blob/master/scripts/irods/setup_options.py#L22

And perhaps maybe even *not* adding the SSL options to the manual prompt
as it might break other people's setup if we ever push upstrem. SSL
options would only be able to be set via the json.

[DHS-1044, DHS-1889]