Use exec form of CMD for ires, ires-s3 & ires-centos
Why?

To speed up `docker-compose down` or `docker-compose stop`.. If
dockerize gets the stop signal (SIGTERM), it will act accordingly and
stop the process. But if the shell[1] gets the signal (result of having CMD
in "shell form"), the signal will be ignored (also not forwarded to
dockerize). 10 seconds later docker will force stop the container.

See: https://github.com/jwilder/dockerize/issues/12
     https://docs.docker.com/engine/reference/builder/#cmd

Note that as centos has bash for /bin/sh, this is not necessary. But
it's converted anyway for consistency. See comment at the end of its
Dockerfile.

Also note that for iCAT, avoiding this 10 second timeout is done
differently, as it doesn't run dockerize. See comments in its
bootstrap.sh for more details (irods/icat/scripts/bootstrap.sh)

[1]: As long as the shell is 'dash' it seems, as it's the case for
     ubuntu images. Bash behaves differently (centos)

[DHDO-163]