Add debug mechanism for iRODS bootstrap scripts
You can set:
    DEBUG_DH_BOOTSTRAP: yes

Note: At the moment this won't work for most bootstrap_irods.sh as they
are called with `su - `, clearing the environment (so
DEBUG_DH_BOOTSTRAP) doesn't make it through). We can either just do `su`
(no `-`), or use sudo. But we might remove sudo later on. So for now,
this is a start.