Fix: circumvent race-condition with project from iCAT and iRES
We expect icat to have created hardcoded projects P000000010 and
P000000011 before we task ires's bootstrap_irods.sh with creating other
projects with test_create_new_project, but we can't be sure P0..10 and
P0..11 have been created yet. So, just in case, we make sure that
test_create_new_project.r (-> create_new_project.py) thinks that the
latest project created is project 11 by setting the
latest_project_number AVU to 11.

Keeping in mind this is *not nice* and error prone. Containers
interdependencies is still a problem to solve.