Add iCAT docker health check
We are experimenting with this feature.

We pretend to use them to control start up sequence, but they are
limited and probably not really meant for that purpose.

`depends_on` health status does not seem to be supported by version 3 of
docker-compose, but it is supported by the now recommended Compose Spec.