Make ires start conditional on healthy icat
This again, has its caveats. Health checks are scheduled every 10
seconds for icat. So, for example, it could happen that icat has been
ready for 9 seconds, and then a health check is executed, bringing the
service (icat) to "healthy". And at that point, ires will be started.
So 9 seconds were wasted.

Ultimately, docker health checks are not very flexible and not meant for
sequence starts. That's what container orchestration tools are for. But
as we agreed, kubernetes is overkill for us at the moment.

[DHDO-426]