Add .env variables for json_configuration-ready dh-irods
Some variables here are just test passwords for development purposes
only. They are equivalent to the setup_responses ones we had here for
example:
https://github.com/MaastrichtUniversity/docker-dev/blob/v3.20.1/icat/setup_responses

But given new constraints imposed by dh-irods now, we change them here.
And add the `test_` prefix to make it clear these are not real passwords
and should never be spotted anywhere that is not test/dev (maybe this is
not a sensible idea though?)

[DHDO-777]